fix(models): stock_prices/stock_features volume 改為 BigInteger
SQLite INTEGER 最大 64-bit，PostgreSQL Integer 只有 32-bit。
台積電成交量可達數十億，超出上限導致遷移失敗。

+++ backend/app/models/stock_feature.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Date, Index
+from sqlalchemy import Column, Integer, BigInteger, String, Float, Date, Index
 from app.db.database import Base
 
 
@@ -47,7 +47,7 @@ class StockFeature(Base):
     bb_pctb = Column(Float)          # %B 位置 (0~1)
 
     # --- 成交量 ---
-    volume = Column(Integer)
+    volume = Column(BigInteger)
     vol_ma5 = Column(Float)          # 5 日均量
     vol_ratio = Column(Float)        # 量比 (volume / vol_ma5)
